// types.ts
export interface WeatherData {
  current: {
    temp: number;
    description: string;
    icon: string;
    humidity: number;
    wind: number;
    location: string;
  };
  daily: {
    date: string;
    day: string;
    min: number;
    max: number;
    icon: string;
    description: string;
  }[];
}